Use Window.Invalidate, not Window.DoRender
The DoRender was only public for a short time by accident in the
engine -- it was not supposed to be used by game code, and it was
fixed to be private at some point. It is unportable and call
"DoRender" from any point you want. Instead, call Window.Invalidate
that is nice (schedules redraw at nearest suitable time), and portable.